java/Computer_Simulation/src/backend/requirements.txt
fastapi
jpype1
uvicorn[standard]